r/storage_blob: moving the legacy sdk functions into their own file to aid refactoring
